Import incomplete example code
This here is a small sample project to demonstrate use of
wireguard-windows' embeddable-dll-service:

https://git.zx2c4.com/wireguard-windows/about/embeddable-dll-service/README.md

It requires tunnel.dll to be built from there. It also requires
wintun.dll, which can be downloaded from https://www.wintun.net/.

This works for logging into Azire and starting the VPN and such, but it
also is very incomplete, potentially insecure, and really made as a
basis of future code rather than something that's production-ready.
Currently it has some limitations:

- Permissions and DACLs have not been worked out yet.
- The build settings hardcode x64 everywhere, which might not be so
  nice.
- The UI is an oldschool winforms UI. It might be desirable to rewrite
  this using UWP or something nicer.
- Does not use DPAPI.

The code in this commit is provided under GPLv2. The motivation for
writing it was to demonstrate embeddable-dll-service' capabilities and
ensure that the underlying code for it hasn't bitrotted, after the
former known-user of the API seemingly abandoned their codebase.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>